Add notary-common crate.
