+++ crates/notary/common/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "notary-common"
+version = "0.1.0-alpha.11-pre"
+description = "Common code shared between notary-server and notary-client"
+edition = "2021"
+
+[lints]
+workspace = true
+
+[dependencies]
+serde = { workspace = true, features = ["derive"] }
