+++ crates/notary/server/src/types.rs
@@ -25,41 +25,12 @@ pub struct InfoResponse {
     pub quote: Quote,
 }
 
-/// Response object of the /session API
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct NotarizationSessionResponse {
-    /// Unique session id that is generated by notary and shared to prover
-    pub session_id: String,
-}
-
-/// Request object of the /session API
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct NotarizationSessionRequest {
-    pub client_type: ClientType,
-    /// Maximum data that can be sent by the prover
-    pub max_sent_data: Option<usize>,
-    /// Maximum data that can be received by the prover
-    pub max_recv_data: Option<usize>,
-}
-
 /// Request query of the /notarize API
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct NotarizationRequestQuery {
     /// Session id that is returned from /session API
     pub session_id: String,
-}
-
-/// Types of client that the prover is using
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub enum ClientType {
-    /// Client that has access to the transport layer
-    Tcp,
-    /// Client that cannot directly access transport layer, e.g. browser
-    /// extension
-    Websocket,
 }
 
 /// Global data that needs to be shared with the axum handlers
